workers/azure-workload-worker/requirements.txt
azure-identity>=1.17.0
azure-mgmt-compute>=31.0.0
azure-mgmt-sql>=3.0.1
azure-mgmt-rdbms>=10.1.1
azure-mgmt-resource>=23.1.0
azure-mgmt-storage>=21.2.0
azure-storage-blob>=12.22.0
aio-pika>=9.4.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
httpx>=0.27.0
six>=1.16.0
fastapi>=0.100.0
python-jose[cryptography]>=3.3.0
